No manager, no rules
