--- app/search/-components/facets-manager/styles.scss
@@ -1,9 +1,0 @@
-.active-filter-list {
-    padding-left: 0;
-    list-style: none;
-}
-
-.active-filter-item {
-    display: flex;
-    justify-content: space-between;
-}
