--- app/search/-components/facets-manager/template.hbs
@@ -1,33 +1,0 @@
-{{#if @activeFilters.length}}
-    <ul local-class='active-filter-list'>
-        {{#each @activeFilters as |filter|}}
-            <li local-class='active-filter-item'>
-                <span>
-                    {{filter.key}} |
-                    {{filter.value}}
-                </span>
-                <Button
-                    aria-label={{t 'search.remove_filter' key=filter.key value=filter.value}}
-                    @layout='fake-link'
-                    {{on 'click' (fn this.toggleFilter filter)}}
-                >
-                    <FaIcon @icon='times' />
-                </Button>
-            </li>
-        {{/each}}
-    </ul>
-{{/if}}
-
-{{#each this.filterableProperties as |filterableProperty|}}
-    {{#let filterableProperty.metadataRecord as |propertyRecord|}}
-        <Search::-Components::FilterFacet
-            @recordSearchText={{@recordSearchText}}
-            @recordSearchFilters={{@activeFilters}}
-            @propertyRecord={{propertyRecord}}
-            @propertySearch={{filterableProperty}}
-            @toggleFilter={{this.toggleFilter}}
-        />
-    {{/let}}
-{{else}}
-    {{t 'search.left_panel.no_filterable_properties'}}
-{{/each}}
